don't bother retrying

+++ cmip6_download.py
@@ -1,30 +1,30 @@
-from concurrent.futures import ThreadPoolExecutor
-from concurrent.futures import ProcessPoolExecutor
 from concurrent.futures import as_completed
-import zipfile
+from concurrent.futures import ProcessPoolExecutor
 import argparse
-import random
-import logging
 import collections
+import logging
+import multiprocessing
 import os
+import pickle
+import random
 import requests
 import shutil
 import sys
-from functools import partial
-from threading import Lock
 import time
+import zipfile
 
-from tenacity import retry, stop_after_attempt, wait_random_exponential
 from rasterio.transform import Affine
+from scipy.interpolate import griddata
 import numpy
+import pandas
 import rasterio
 import xarray
-import pandas
 
 BASE_URL = 'https://esgf-node.llnl.gov/esg-search/search'
 BASE_SEARCH_URL = 'https://esgf-node.llnl.gov/search_files'
 VARIANT_SUFFIX = 'i1p1f1'
 LOCAL_CACHE_DIR = '_cmip6_local_cache'
+PROCESSED_FILES_PICKLE = os.path.join(LOCAL_CACHE_DIR, 'processed_files.pkl')
 HOT_DIR = 'D:/hot_cache'
 for dir_path in [LOCAL_CACHE_DIR, HOT_DIR]:
     os.makedirs(dir_path, exist_ok=True)
@@ -40,6 +40,29 @@ LOGGER.setLevel(logging.DEBUG)
 logging.getLogger('fetch_data').setLevel(logging.INFO)
 
 
+class ProcessedFiles:
+    def __init__(self, pickle_path, manager):
+        self.lock = manager.Lock()
+        self.pickle_path = pickle_path
+        if os.path.exists(self.pickle_path):
+            LOGGER.debug(f'loading from {self.pickle_path}')
+            with open(self.pickle_path, 'rb') as f:
+                self.set = pickle.load(f)
+                LOGGER.debug(self.set)
+        else:
+            self.set = set()
+
+    def add(self, file):
+        with self.lock:
+            self.set.add(file)
+            with open(self.pickle_path, 'wb') as f:
+                pickle.dump(self.set, f)
+
+    def __contains__(self, file):
+        with self.lock:
+            return file in self.set
+
+
 def handle_retry_error(retry_state):
     # retry_state.outcome is a built-in tenacity method that contains the result or exception information from the last call
     last_exception = retry_state.outcome.exception()
@@ -53,11 +76,19 @@ def handle_retry_error(retry_state):
             str(last_exception.statistics).replace('\n', '<enter>')+"\n")
 
 
-def _download_and_process_file(args):
+def _download_and_process_file(processed_files, args):
     try:
+        local_zip_path = None
+        netcdf_path = None
         LOGGER.info(f'********* processing {args}')
         variable, scenario, model, variant, url = args
+        LOGGER.debug(f'checking {url}')
+        if url in processed_files:
+            return True
         netcdf_path = _download_file(HOT_DIR, url)
+        if not os.path.exists(netcdf_path):
+            raise ValueError(
+                f'expected a file at {netcdf_path} but nothing was there downloaded from {url}')
         base_path_pattern = (
             r'cmip6/{variable}/{scenario}/{model}/{variant}/'
             r'cmip6-{variable}-{scenario}-{model}-{variant}-{date}.tif')
@@ -81,12 +112,16 @@ def _download_and_process_file(args):
                 local_zip_path = os.path.join(HOT_DIR, zip_path_pattern)
                 target_zip_path = os.path.join(
                     LOCAL_CACHE_DIR, zip_path_pattern)
-                zip_files(file_list, local_zip_path, target_zip_path)
+                if not os.path.exists(target_zip_path):
+                    zip_files(file_list, local_zip_path, target_zip_path)
             LOGGER.info(f'done processing {os.path.basename(url)}')
         hot_dir = os.path.dirname(local_zip_path)
         LOGGER.info(f'removing directory {os.path.dirname(hot_dir)}')
-        shutil.rmtree(os.path.dirname(local_zip_path))
-        os.remove(netcdf_path)
+        processed_files.add(url)
+        if local_zip_path and os.path.exists(local_zip_path):
+            shutil.rmtree(os.path.dirname(local_zip_path))
+        if netcdf_path and os.path.exists(netcdf_path):
+            os.remove(netcdf_path)
         return True
     except Exception:
         LOGGER.exception(f'error on _download_and_process_file {args}')
@@ -101,9 +136,6 @@ def zip_files(file_list, local_path, target_path):
     shutil.move(local_path, target_path)
 
 
-@retry(stop=stop_after_attempt(5),
-       wait=wait_random_exponential(multiplier=1, min=1, max=10),
-       retry_error_callback=handle_retry_error)
 def _download_file(target_dir, url):
     try:
         stream_path = os.path.join(
@@ -144,6 +176,10 @@ def _download_file(target_dir, url):
     except Exception:
         LOGGER.exception(
             f'failed to download {url} to {target_dir}, possibly retrying')
+        if os.path.exists(stream_path):
+            os.remvoe(stream_path)
+        if os.path.exists(target_path):
+            os.remvoe(target_path)
         raise
 
 
@@ -194,6 +230,7 @@ def process_cmip6_netcdf_to_geotiff(
                 previous_year = year
             data_values = daily_data.values[numpy.newaxis, ...]
             coord_list = []
+            coord_id_list = []
             res_list = []
             for coord_id, field_options in zip(['x', 'y'], [
                     ['longitude', 'long', 'lon'],
@@ -205,11 +242,33 @@ def process_cmip6_netcdf_to_geotiff(
                             (coord_array[-1] - coord_array[0]) /
                             len(coord_array)))
                         coord_list.append(coord_array)
+                        coord_id_list.append(field_id)
+                        break
                     except KeyError:
                         pass
             if len(coord_list) != 2:
                 raise ValueError(
                     f'coord list not fully defined for {netcdf_path}')
+            if len(daily_data.values.shape) == 1:
+                # this is 1D data, need to re-create 2D grid
+                lon = coord_list[0].values
+                lat = coord_list[1].values
+                # Define the grid
+                grid_size = 0.25
+                n_lat = int(numpy.round((lat.max()-lat.min())/grid_size))
+                n_lon = int(numpy.round((lon.max()-lon.min())/grid_size))
+                lats = numpy.linspace(lat.min(), lat.max(), n_lat)
+                lons = numpy.linspace(lon.min(), lon.max(), n_lon)
+                lon_grid, lat_grid = numpy.meshgrid(lons, lats)
+
+                res_list = [grid_size, grid_size]
+                coord_list = [lons, lats]
+
+                # Perform the 2D interpolation
+                data_values = griddata(
+                    (lon, lat), daily_data, (lon_grid, lat_grid),
+                    method='linear')
+                data_values = data_values[numpy.newaxis, ...]
             transform = Affine.translation(
                 *[a[0] for a in coord_list]) * Affine.scale(*res_list)
 
@@ -233,6 +292,8 @@ def process_cmip6_netcdf_to_geotiff(
         return raster_by_year
     except Exception:
         LOGGER.exception(f'error on {netcdf_path}')
+        raise
+
 
 
 def _write_raster(data_values, coord_list, transform, target_path):
@@ -266,32 +327,34 @@ def main():
         'Process CMIP6 raw urls to geotiff.'))
     parser.add_argument('url_list_path', help='url list')
     args = parser.parse_args()
+    with multiprocessing.Manager() as manager:
+        processed_files = ProcessedFiles(PROCESSED_FILES_PICKLE, manager)
 
-    # pr,historical,SAM0-UNICON,r1i1p1f1,http://aims3.llnl.gov/thredds/fileServer/css03_data/CMIP6/CMIP/SNU/SAM0-UNICON/historical/r1i1p1f1/day/pr/gn/v20190323/pr_day_SAM0-UNICON_historical_r1i1p1f1_gn_19000101-19001231.nc
+        # pr,historical,SAM0-UNICON,r1i1p1f1,http://aims3.llnl.gov/thredds/fileServer/css03_data/CMIP6/CMIP/SNU/SAM0-UNICON/historical/r1i1p1f1/day/pr/gn/v20190323/pr_day_SAM0-UNICON_historical_r1i1p1f1_gn_19000101-19001231.nc
 
-    with open(args.url_list_path, 'r') as file:
-        param_and_url_list = [line.rstrip().split(',') for line in file]
+        with open(args.url_list_path, 'r') as file:
+            param_and_url_list = [line.rstrip().split(',') for line in file]
 
-    random.seed(1)
-    random.shuffle(param_and_url_list)
-    # for index, val in enumerate(param_and_url_list):
-    #     print(f'{index}: {val}')
-    # return
-    #param_and_url_list = [param_and_url_list[2]]
-    start_time = time.time()
-    with ProcessPoolExecutor(5) as global_executor:
-        future_list = {
-            global_executor.submit(
-                _download_and_process_file, param_and_url_arg)
-            for param_and_url_arg in param_and_url_list}
+        random.seed(1)
+        random.shuffle(param_and_url_list)
+        # for index, val in enumerate(param_and_url_list):
+        #     print(f'{index}: {val}')
+        # return
+        #param_and_url_list = [param_and_url_list[2]]
+        start_time = time.time()
+        with ProcessPoolExecutor(5) as global_executor:
+            future_list = {
+                global_executor.submit(
+                    _download_and_process_file, processed_files, param_and_url_arg)
+                for param_and_url_arg in param_and_url_list[0:5]}
 
-        for future in as_completed(future_list):
-            try:
-                _ = future.result()  # This will raise an exception if the worker function failed
-            except Exception:
-                LOGGER.exception('something failed on download and process')
-                global_executor.shutdown(wait=False)
-                raise  # Re-raise the original exception
+            for future in as_completed(future_list):
+                try:
+                    _ = future.result()  # This will raise an exception if the worker function failed
+                except Exception:
+                    LOGGER.exception('something failed on download and process')
+                    global_executor.shutdown(wait=False)
+                    raise  # Re-raise the original exception
     LOGGER.info(f'all done took {time.time()-start_time:.2f}s')
 
 if __name__ == '__main__':
